`react-server` as a separate plugin.
