+++ src/react-server/test/.eslintrc
@@ -1,0 +1,6 @@
+{
+  "rules": {
+    "func-names": 0,
+    "import/no-extraneous-dependencies": 0
+  }
+}
