--- src/react/server/index.js
@@ -1,5 +1,0 @@
-import renderToString from './renderToString';
-
-export default {
-  renderToString,
-};
